- Loading system variables if .env file does not exists. - Removed code related to video_chunks table, also added auto migrate logic for GORM models and removed database.sql for manual db migration.

+++ internal/grpc/server/video_catalog.go
@@ -72,7 +72,7 @@ func (v *videoCatalogServer) FindById(ctx context.Context, data *videocatalogpb.
 
 	err := database.Conn.
 		Select("id", "title", "description", "thumbnail", "path", "published_at", "duration", "resolution").
-		Where(&model.Video{Id: uint(data.Id)}).
+		Where(&model.Video{Id: uint64(data.Id)}).
 		First(&video).Error
 
 	if err != nil {
